These should do the trick
ON DELETE CASCADE is essential for proper database functionality. Otherwise, to delete a movie, you first must delete every instance of an actor acting in it, then any instance of someone renting it, then any awards it's won, then the Title it refers to, then the movie itself. Who's got time for that?